requirements.txt
fastapi
uvicorn
sse-starlette
redis
python-dotenv
google-genai
mss
pyautogui
pywinauto
pytesseract
easyocr
opencv-python
numpy
pillow
pypiwin32